+++ tests/testthat/test-point_metrics.R
@@ -13,19 +13,19 @@ test_that("points_metrics works with a single metric (knn)", {
   m = point_metrics(las, ~mean(Z), k = 3L)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 4))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$V1, c(0.5, 1, 0.5, 1.5, 1, 1.5))
-  expect_equal(names(m), c("X", "Y", "Z", "V1"))
+  expect_equal(names(m), c("pointID", "V1"))
 
   m = point_metrics(las, ~list(M = mean(Z)), k = 3L)
 
-  expect_equal(dim(m), c(npoints(las), 4))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$M, c(0.5, 1, 0.5, 1.5, 1, 1.5))
+  expect_equal(names(m), c("pointID", "M"))
+
+  m = point_metrics(las, ~list(M = mean(Z)), k = 3L, xyz = TRUE)
+
   expect_equal(names(m), c("X", "Y", "Z", "M"))
-
-  m = point_metrics(las, ~list(M = mean(Z)), k = 3L, xyz = FALSE)
-
-  expect_equal(names(m), c("M"))
 })
 
 test_that("points_metrics works with a single metric (shpere)", {
@@ -33,19 +33,19 @@ test_that("points_metrics works with a single metric (shpere)", {
   m = point_metrics(las, ~length(Z), r = 0.8)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 4))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$V1, c(2,2,3,3,2,2))
-  expect_equal(names(m), c("X", "Y", "Z", "V1"))
+  expect_equal(names(m), c("pointID", "V1"))
 
   m = point_metrics(las, ~list(M = mean(Z)), r = 0.8)
 
-  expect_equal(dim(m), c(npoints(las), 4))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$M, c(0.25, 1.25, 0.5, 1.5, 0.75, 1.75))
+  expect_equal(names(m), c("pointID", "M"))
+
+  m = point_metrics(las, ~list(M = mean(Z)), r = 3L, xyz = TRUE)
+
   expect_equal(names(m), c("X", "Y", "Z", "M"))
-
-  m = point_metrics(las, ~list(M = mean(Z)), r = 3L, xyz = FALSE)
-
-  expect_equal(names(m), c("M"))
 })
 
 test_that("points_metrics restpect the filter argument (knn)", {
@@ -53,7 +53,7 @@ test_that("points_metrics restpect the filter argument (knn)", {
   m = point_metrics(las, ~mean(Z), k = 3L, filter = ~I>2)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las)-2, 4))
+  expect_equal(dim(m), c(npoints(las)-2, 2))
   expect_equal(m$V1, c(1, 1.5, 1, 1.5))
 })
 
@@ -62,7 +62,7 @@ test_that("points_metrics restpect the filter argument (shpere)", {
   m = point_metrics(las, ~mean(Z), r = 0.8, filter = ~I>2)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las)-2, 4))
+  expect_equal(dim(m), c(npoints(las)-2, 2))
   expect_equal(m$V1, c(0.75, 1.75, 0.75, 1.75))
 })
 
@@ -71,12 +71,12 @@ test_that("points_metrics works with a multiple metrics (knn)", {
   m = point_metrics(las, ~list(mean(Z), max(Z), Z[1]), k = 3L)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 6))
+  expect_equal(dim(m), c(npoints(las), 4))
   expect_equal(m$V3, c(0, 1, 0.5, 1.5, 1, 2))
 
-  m = point_metrics(las, ~list(A = mean(Z), B = max(Z), C = Z[1]), k = 3L, xyz = FALSE)
+  m = point_metrics(las, ~list(A = mean(Z), B = max(Z), C = Z[1]), k = 3L, xyz = TRUE)
 
-  expect_equal(names(m), c("A", "B", "C"))
+  expect_equal(names(m), c("X", "Y", "Z", "A", "B", "C"))
 })
 
 test_that("points_metrics works with a multiple metrics (sphere)", {
@@ -84,12 +84,12 @@ test_that("points_metrics works with a multiple metrics (sphere)", {
   m = point_metrics(las, ~list(mean(Z), max(Z), Z[1]), r = 0.8)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 6))
+  expect_equal(dim(m), c(npoints(las), 4))
   expect_equal(m$V1, c(0.25, 1.25, 0.5, 1.5, 0.75, 1.75))
 
-  m = point_metrics(las, ~list(A = mean(Z), B = max(Z), C = Z[1]), r = 0.8, xyz = FALSE)
+  m = point_metrics(las, ~list(A = mean(Z), B = max(Z), C = Z[1]), r = 0.8, xyz = TRUE)
 
-  expect_equal(names(m), c("A", "B", "C"))
+  expect_equal(names(m), c("X", "Y", "Z", "A", "B", "C"))
 })
 
 
@@ -98,12 +98,12 @@ test_that("points_metrics works with lidR metrics", {
   m = point_metrics(las, .stdmetrics_z, k = 3L, xyz = FALSE)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 36))
+  expect_equal(dim(m), c(npoints(las), 37))
 
   m = point_metrics(las, .stdmetrics_z, r = 0.8, xyz = FALSE)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 36))
+  expect_equal(dim(m), c(npoints(las), 37))
 })
 
 test_that("points_metrics works with nested function", {
@@ -113,7 +113,7 @@ test_that("points_metrics works with nested function", {
   m <- g(las)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 1))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$V1, c(1, 1.5, 1, 2, 1.5, 2))
 
   f <- function(x, y, z) {  return(max(c(x,y,z))) }
@@ -121,7 +121,7 @@ test_that("points_metrics works with nested function", {
   m <- g(las)
 
   expect_is(m, "data.table")
-  expect_equal(dim(m), c(npoints(las), 1))
+  expect_equal(dim(m), c(npoints(las), 2))
   expect_equal(m$V1, c(0.5, 1.5, 1, 2, 1, 2))
 })
 
@@ -158,7 +158,7 @@ test_that("points_metrics fails", {
 
 test_that("points_metrics return references on coordinates", {
 
-  m = point_metrics(las, ~list(mean(Z), max(Z), Z[1]), k = 3L)
+  m = point_metrics(las, ~list(mean(Z), max(Z), Z[1]), k = 3L, xyz = TRUE)
 
   expect_reference(m$X, las$X)
   expect_reference(m$Y, las$Y)